+++ src/resources.rs
@@ -1,12 +1,37 @@
 use crate::*;
 use bevy_rapier3d::prelude::Collider;
+use bevy_renet::renet::NETCODE_USER_DATA_BYTES;
 use clap::Parser;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
+#[derive(Serialize, Deserialize, Component, Debug, Copy, Clone)]
+pub struct ClientPreferences {
+    pub autostart: bool,
+}
+
+#[derive(Serialize, Deserialize, Component, Debug, Copy, Clone)]
+pub struct ClientData {
+    pub preferences: ClientPreferences,
+    pub inhabited_mass_id: u64,
+}
+
+impl ClientPreferences {
+    pub fn to_netcode_user_data(self) -> [u8; NETCODE_USER_DATA_BYTES] {
+        let mut user_data = [0u8; NETCODE_USER_DATA_BYTES];
+        user_data[0] = self.autostart as u8;
+        user_data
+    }
+
+    pub fn from_user_data(user_data: &[u8; NETCODE_USER_DATA_BYTES]) -> Self {
+        let autostart = user_data[0] == 1_u8;
+        Self { autostart }
+    }
+}
+
 #[derive(Default, Resource, Debug)]
 pub struct Lobby {
-    pub clients: HashMap<u64, wat::ClientData>,
+    pub clients: HashMap<u64, ClientData>,
 }
 
 #[derive(Parser, Resource)]
